Stub out the demo on Windows
